+++ docs/source/api/misc.rst
@@ -7,5 +7,6 @@ Other utils
    :toctree: generated/
 
    compute_log_likelihood
+   compute_log_prior
    find_constrained_prior
    DictToArrayBijection
